requirements.txt
# Core Data Engineering
pandas==2.2.2
numpy==1.26.4
jsonschema==4.18.0      # Explicitly pinned to compatible version
# apache-airflow==2.9.1   # Removed due to SQLAlchemy conflicts
prefect==3.4.14         # Orchestration
sqlalchemy==2.0.29      # Explicitly set to compatible version
kafka==1.3.5            # Stream processing
# kafka-python==2.0.2   # Alternative stream processing
kafka-python-ng==2.2.3  # Alternative stream processing
minio==7.2.4            # S3-compatible storage
psycopg2-binary==2.9.9  # PostgreSQL adapter
timescaledb==0.0.4      # Time-series extension

# Machine Learning
scikit-learn==1.4.2
xgboost==2.0.3
lightgbm==4.3.0
catboost==1.2.7
mlflow==2.13.0         # Model tracking
# seldon-core==2.9.0     # Model serving

# Deep Learning
# torch==2.2.2
# tensorflow==2.16.1

# APIs & Web Services
fastapi==0.111.0
uvicorn==0.28.0        # Downgraded to compatible version (NOT 0.29.0)
requests==2.31.0
boto3==1.28.0

# Geospatial & Routing
geopy==2.4.1
osmnx==1.9.1           # OpenStreetMap routing
valhalla==0.1.4        # Alternative routing engine
geopandas==0.14.4

# Visualization
plotly==5.22.0
dash==2.16.1           # Dashboard framework
grafana-client==4.3.1  # For Grafana integration

# Workflow & Infrastructure
docker==7.0.0
kubernetes==29.0.0
dask==2024.5.1  # Parallel processing
faust-streaming[rocksdb]==0.11.3   # Streaming application

# Utilities
python-dotenv==1.0.1   # Environment management
pyyaml==6.0.1          # Configuration files
click==8.1.7           # Latest compatible version for dask and prefect
referencing==0.36.2